% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funs.R
\name{age.to.length}
\alias{age.to.length}
\title{Convert fish age to length using von Bertalanffy 3-param equation}
\usage{
age.to.length(N, age, year, linf, k, t0, cvdist)
}
\arguments{
\item{N}{matrix with numbers at age}

\item{age}{max age}

\item{year}{current year in simulation}

\item{linf}{L-infinity value}

\item{k}{K coefficient}

\item{t0}{age at length 0}

\item{cvdist}{vector of CVs to sample from}
}
\description{
Convert fish age to length using von Bertalanffy 3-param equation
}
\examples{
age.to.length()
}
\keyword{Bertalanffy}
\keyword{age,}
\keyword{length,}
\keyword{von}
