% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_to_length.R
\name{age_to_len}
\alias{age_to_len}
\title{Convert numbers at age to numbers at length}
\usage{
age_to_len(ALK, Nage, year, age_0 = T)
}
\arguments{
\item{ALK}{age-length key from SS report}

\item{Nage}{dataframe with N-at-age}

\item{year}{current year in simulation}

\item{age_0}{include age 0 in numbers at age matrix, default is true}
}
\description{
Convert numbers at age to numbers at length
}
