% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funs.R
\name{catch.in.biomass}
\alias{catch.in.biomass}
\title{Converts catch in numbers to catch in biomass}
\usage{
catch.in.biomass(dat.list, N, year, z, f.by.fleet)
}
\arguments{
\item{dat.list}{Contains selectivity-at-age matrix and weight-at-age matrix}

\item{N}{matrix with numbers at age}

\item{year}{current year in simulation}

\item{z}{total mortality at age vector}

\item{f.by.fleet}{Fishing mortality by fleet for year y}
}
\description{
Converts catch in numbers to catch in biomass
}
\examples{


}
\keyword{catch}
