% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funs.R
\name{catch.in.number}
\alias{catch.in.number}
\title{Calculates numbers at age of catch}
\usage{
catch.in.number(fsel, N, year, z, f.by.fleet, se)
}
\arguments{
\item{fsel}{Selectivity-at-age matrix of fisheries}

\item{N}{matrix with numbers at age}

\item{year}{current year in simulation}

\item{z}{total mortality at age vector}

\item{f.by.fleet}{Fishing mortality by fleet for year y}

\item{se}{vector of log(catch error)}
}
\description{
Calculates numbers at age of catch
}
\examples{
catch.in.number()

}
\keyword{catch}
