% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence_funs.R
\name{check_convergence}
\alias{check_convergence}
\title{Check for SS convergence}
\usage{
check_convergence(dir.)
}
\arguments{
\item{dir.}{file path for assessment files}
}
\value{
a vector of values: the number of parameters, the objective function value, and the max gradient component
}
\description{
Check for SS convergence
}
\keyword{convergence}
