% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competition_fun.R
\name{competition_fun}
\alias{competition_fun}
\title{Convert numbers at age to numbers at length}
\usage{
competition_fun(ALK, year, N, Nrs, ALK_r, beta.vec, carry.K)
}
\arguments{
\item{ALK}{age-length key from SS report}

\item{year}{current year in simulation}

\item{N}{dataframe with N-at-age for species 1 (vermilion)}

\item{Nrs}{dataframe with N-at-age for species 2 (red)}

\item{ALK_r}{age-length key for red snapper or (species 2)}

\item{beta.vec}{vector of length-specific betas from TMB model}

\item{carry.K}{combined carrying capacity}
}
\description{
Convert numbers at age to numbers at length
}
