% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_funs.R
\name{copy_files}
\alias{copy_files}
\title{Copy assessment files to a new folder}
\usage{
copy_files(dat.list, dir., new.dir)
}
\arguments{
\item{dat.list}{a list object containing the sequence of years}

\item{dir.}{current directory for scenario}

\item{new.dir}{the new destitnation to send files to}
}
\description{
Copy assessment files to a new folder
}
\keyword{assessment}
\keyword{files}
