% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_dat.R
\name{dat.update}
\alias{dat.update}
\title{This takes simulated data from the operating model pop dynamics and adds it in the correct format to the .dat file for SS.}
\usage{
dat.update(year, dat.list, dat., agecomp.list, I, .datcatch, comp.I, dir.,
  write = T)
}
\arguments{
\item{year}{last year in sequence of new data}

\item{dat.list}{list that has the sequence of years}

\item{agecomp.list}{simulated age comps from catch}

\item{I}{simulated indices of abundance for each fleet}

\item{.datcatch}{A dataframe with the catch for each fishing fleet in the correct units (biomass and numbers)}

\item{comp.I}{index of red snapper abundance, a measure of competition}

\item{dir.}{directory of where to read and write new files from/to}

\item{write}{default set to TRUE, if FALSE, new .dat file will not be saved (won't overwrite the old file)}

\item{.dat}{data file that needs to be updated}
}
\description{
This takes simulated data from the operating model pop dynamics and adds it in the correct format to the .dat file for SS.
}
\examples{

}
\keyword{SS}
\keyword{data,}
\keyword{update}
