% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funs.R
\name{exploit_catch}
\alias{exploit_catch}
\title{Calculates catch-at-age by fleet based on exploitation rates}
\usage{
exploit_catch(F_exploit, N, year, dat.list)
}
\arguments{
\item{F_exploit}{a vector of exploitation rates for each fishery}

\item{N}{matrix with numbers at age}

\item{year}{current year in simulation}

\item{dat.list}{Contains selectivity-at-age matrix and weight-at-age matrix}
}
\description{
Calculates catch-at-age by fleet based on exploitation rates
}
\keyword{catch}
\keyword{exploitation}
