% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_for_F.R
\name{H_rate}
\alias{H_rate}
\title{Function to solve for harvest(exploitation) rate of a fleet based on catch for that year}
\usage{
H_rate(catch, fleet, N, year, dat.list, total.catch = F)
}
\arguments{
\item{catch}{the expected catch from fleet for the year (in units of the fleet)}

\item{fleet}{fleet you are calculating F for}

\item{N}{matrix of numbers-at-age, first column is year}

\item{year}{year index}

\item{dat.list}{list including M, Nages, weight-at-age, and selectivity}

\item{total.catch}{is the projected catch a combined catch (T) or only for that fleet (F)}
}
\value{
The harvest rate that achieves the projected catch level
}
\description{
Function to solve for harvest(exploitation) rate of a fleet based on catch for that year
}
\keyword{harvest}
\keyword{projections}
\keyword{rate,}
