% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence_funs.R
\name{jit_for_converg}
\alias{jit_for_converg}
\title{Run jitter analysis if model didn't converge}
\usage{
jit_for_converg(convCheck, dir., lin = FALSE)
}
\arguments{
\item{convCheck}{vector of convergence values (from check_convergence)}

\item{dir.}{file path for assessment files}

\item{lin}{if TRUE running on a linux system, default is FALSE}
}
\value{
a logical term (T or F), T if the model converged from the jitter, F if the model didn't converge within 3 tries
}
\description{
Run jitter analysis if model didn't converge
}
\keyword{convergence,}
\keyword{jitter}
