% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funs.R
\name{num_to_bio}
\alias{num_to_bio}
\title{Converts numbers-at-age to biomass-at-age or vice versa.
Inputs need to be a dataframe, not a matrix.}
\usage{
num_to_bio(df, dat.list, Natage = T, age0 = F)
}
\arguments{
\item{df}{dataframe of values you are converting}

\item{dat.list}{list that contains matrix of weights at age}

\item{Natage}{default is TRUE, T converts N-at-age to biomass, F converts B-at-age to numbers}

\item{age0}{defualt is FALSE, will exclude age0 in calculations, TRUE will include them}
}
\description{
Converts numbers-at-age to biomass-at-age or vice versa.
Inputs need to be a dataframe, not a matrix.
}
\examples{


}
\keyword{biomass-at-age}
\keyword{numbers-at-age,}
