% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_funs.R
\name{Numextract}
\alias{Numextract}
\title{Extract multiple numbers from a character string}
\usage{
Numextract(string)
}
\arguments{
\item{string}{the character string containing numbers to extract}
}
\description{
Extract multiple numbers from a character string
}
\keyword{character}
\keyword{numbers,}
\keyword{string}
