% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebuild.R
\name{p_star}
\alias{p_star}
\title{Find the acceptable probability of going past the OFL (P*)}
\usage{
p_star(Assess_info = "2", OFL_uncert = "2", Retro_pats = "3",
  Environ = "3", min.risk = 0.3, max.risk = 0.5)
}
\arguments{
\item{Assess_info}{number for how much info is available for assessement}

\item{OFL_uncert}{number describing the level of uncertainty acknowledged in the model}

\item{Retro_pats}{number describing the retrospective patterns}

\item{Environ}{number describing if environmental covariates were included in model}

\item{min.risk}{the level of minimal risk (default at .3)}

\item{max.risk}{the level of max risk (default at .5)}
}
\description{
Find the acceptable probability of going past the OFL (P*)
}
\keyword{p-star}
\keyword{rebuild,}
