% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_funs.R
\name{rand_intvect}
\alias{rand_intvect}
\title{This creates a vector of N whole numbers that add up to M.}
\usage{
rand_intvect(N, M, sd = 1, pos.only = TRUE)
}
\arguments{
\item{N}{number of values in vector}

\item{M}{total values should add up to}

\item{sd}{sd of values, default is 1}

\item{pos.only}{should the vector contain only positive values, default is TRUE}
}
\description{
This creates a vector of N whole numbers that add up to M.
}
\keyword{discrete}
\keyword{random,}
