% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_funs.R
\name{rand_vect_cont}
\alias{rand_vect_cont}
\title{This creates a vector of N numbers that add up to M (can be decimals).}
\usage{
rand_vect_cont(N, M, sd = 1)
}
\arguments{
\item{N}{number of values in vector}

\item{M}{total values should add up to}

\item{sd}{sd of values, default is 1}
}
\description{
This creates a vector of N numbers that add up to M (can be decimals).
}
\examples{
rand_vect_cont()
}
\keyword{continuous}
\keyword{random,}
