% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebuild.R
\name{rebuild_f}
\alias{rebuild_f}
\title{Find the catch and f for time t_target required to rebuild stock}
\usage{
rebuild_f(forefile, dir., dat.list, t_targ)
}
\arguments{
\item{forefile}{full or relative path to forecast file}

\item{dir.}{directory to send new forecast file}

\item{dat.list}{list with the sequence of years}

\item{t_targ}{calculated by rebuild_ttarg, the number of years it will take to rebuild stock}
}
\description{
Find the catch and f for time t_target required to rebuild stock
}
\keyword{F}
\keyword{catch,}
\keyword{rebuild,}
\keyword{t_target,}
