% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ss.R
\name{run_ss}
\alias{run_ss}
\title{This runs SS and jitters if the model does not converge.}
\usage{
run_ss(dir.)
}
\arguments{
\item{dir.}{the directory of where to run ss from}
}
\description{
This runs SS and jitters if the model does not converge.
}
\keyword{SS}
