% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funs.R
\name{simAgecomp}
\alias{simAgecomp}
\title{Generates age composition data from simulated catches}
\usage{
simAgecomp(catch.by.fleet, year, dat.list)
}
\arguments{
\item{catch.by.fleet}{matrix with catches at age by fleet}

\item{year}{current year in simulation}

\item{dat.list}{list with vector of error for each age and vector of the years in simulation}
}
\description{
Generates age composition data from simulated catches
}
\keyword{age}
\keyword{composition}
