% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funs.R
\name{simBatage}
\alias{simBatage}
\title{Calculates biomass at age for fishery dependent surveys}
\usage{
simBatage(N, Flt, asel, year)
}
\arguments{
\item{N}{matrix with numbers at age}

\item{Flt}{max fishing fleet number (3)}

\item{asel}{selectivity-at-age matrix}

\item{year}{current year in simulation}
}
\description{
Calculates biomass at age for fishery dependent surveys
}
\examples{
simBatage()

}
\keyword{biomass-at-age}
