% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funs.R
\name{simBatlen}
\alias{simBatlen}
\title{Calculates biomass at age for fishery independent surveys}
\usage{
simBatlen(Nlen, dat.list, year)
}
\arguments{
\item{Nlen}{matrix with numbers at length}

\item{dat.list}{list with the number of fishery independent surveys and a selectivity-at-length matrix}

\item{year}{current year in simulation}
}
\description{
Calculates biomass at age for fishery independent surveys
}
\keyword{biomass-at-length}
