% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funs.R
\name{simCompetition}
\alias{simCompetition}
\title{Generates competition index based on relative Red Snapper abundances}
\usage{
simCompetition(r, beta, N, sigma, K, Nj, year)
}
\arguments{
\item{r}{rate of population growth}

\item{beta}{competition coeffiecient}

\item{N}{matrix of numbers at age}

\item{sigma}{standard deviation}

\item{K}{carrying capacity}

\item{Nj}{abundance of species j}

\item{year}{current year}
}
\description{
Generates competition index based on relative Red Snapper abundances
}
\keyword{competition,}
\keyword{index}
