% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funs.R
\name{simIndex}
\alias{simIndex}
\title{Generates indices of abundance based on simulated catch}
\usage{
simIndex(q, b, se, year)
}
\arguments{
\item{q}{vector of catchability values for each fleet}

\item{b}{vector of vulnerable biomass for each fleet}

\item{se}{error for each index}

\item{year}{current year in simulation}
}
\description{
Generates indices of abundance based on simulated catch
}
\examples{
simIndex()

}
\keyword{abundance}
\keyword{index}
\keyword{of}
