% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_for_F.R
\name{solve_for_f}
\alias{solve_for_f}
\title{Function to solve for F by fleet based on catch for that year}
\usage{
solve_for_f(proj_catch, fleet, dat.list, year, N)
}
\arguments{
\item{proj_catch}{the expected catch from fleet for the year}

\item{fleet}{fleet you are calculating F for}

\item{dat.list}{list including M, Nages, weight-at-age, and selectivity}

\item{year}{year index}

\item{N}{matrix of numbers-at-age, first column is year}
}
\value{
The F value that achieves the projected catch level
}
\description{
Function to solve for F by fleet based on catch for that year
}
\keyword{F,}
\keyword{projections}
