% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_funs.R
\name{split.recombine}
\alias{split.recombine}
\title{General use functions for data generation and manipulation}
\usage{
\method{split}{recombine}(df1, df2, ind, N)
}
\arguments{
\item{df1}{first dataframe you want to split}

\item{df2}{second dataframe you want to split}

\item{ind}{index attribute you want to group/split dataframes by}

\item{N}{number of individual indices}
}
\description{
This is a generic function to split and recombine dataframes based on common attributes.
I use this for spliting sections of the data files from SS by fleet, adding in the new data, and then recombining everything into the original format.
}
\examples{
split.recombine()
}
\keyword{recombine}
\keyword{split,}
