% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_pub_theme.R
\name{theme_pub}
\alias{theme_pub}
\title{Publication style ggplot theme}
\usage{
theme_pub(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{size font of text}

\item{base_family}{font family (serif, sans)}
}
\description{
Publication style ggplot theme
}
\examples{
theme_pub()

}
