% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funs.R
\name{zatage}
\alias{zatage}
\title{Calcualte total mortality at age}
\usage{
zatage(M, sel, f.by.fleet)
}
\arguments{
\item{M}{natural mortality at age, matrix or df}

\item{sel}{matrix of selectivities at age for each fleet}

\item{f.by.fleet}{Fishing mortality by fleet for year y}
}
\description{
Calcualte total mortality at age
}
\examples{
zatage()

}
\keyword{mortality}
