% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funs.R
\name{catch.in.number}
\alias{catch.in.number}
\title{Calculates numbers at age of catch}
\usage{
catch.in.number(dat.list, N, year, z, f.by.fleet)
}
\arguments{
\item{dat.list}{Contains selectivity-at-age matrix of fisheries and catch SE}

\item{N}{matrix with numbers at age}

\item{year}{current year in simulation}

\item{z}{total mortality at age vector}

\item{f.by.fleet}{Fishing mortality by fleet for year y}
}
\description{
Calculates numbers at age of catch
}
\keyword{catch}
