% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference_points.R
\name{getRP}
\alias{getRP}
\title{Get reference point values after assessment run}
\usage{
getRP(rep., dat.list, year)
}
\arguments{
\item{rep.}{report file from SS assessment}

\item{dat.list}{contains sequence of years in simulation}

\item{year}{current year in simulation}
}
\description{
Get reference point values after assessment run
}
\examples{


}
