% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebuild.R
\name{rebuild_ttarg}
\alias{rebuild_ttarg}
\title{Find the time it takes to rebuild the stock once SSB is below 30\%}
\usage{
rebuild_ttarg(forefile, dir., dat.list)
}
\arguments{
\item{forefile}{full or relative path to forecast file}

\item{dir.}{directory to send new forecast file}

\item{dat.list}{list with the sequence of years}
}
\description{
Find the time it takes to rebuild the stock once SSB is below 30\%
}
\keyword{rebuild,}
\keyword{t_target}
