% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funs.R
\name{simAgecomp}
\alias{simAgecomp}
\title{Generates age composition data from simulated catches}
\usage{
simAgecomp(catch.by.fleet, year, ageerror)
}
\arguments{
\item{catch.by.fleet}{matrix with catches at age by fleet}

\item{year}{current year in simulation}

\item{ageerror}{vector of error for each age}
}
\description{
Generates age composition data from simulated catches
}
\examples{
simAgecomp()

}
\keyword{age}
\keyword{composition}
