% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funs.R
\name{simBatage}
\alias{simBatage}
\title{Calculates biomass at age for fishery dependent surveys}
\usage{
simBatage(N, dat.list, year)
}
\arguments{
\item{N}{matrix with numbers at age}

\item{dat.list}{list with weight-at-age matrix, selectivity-at-age matrix, number of ages, and number of fishing fleets}

\item{year}{current year in simulation}
}
\description{
Calculates biomass at age for fishery dependent surveys
}
\examples{


}
\keyword{biomass-at-age}
