% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funs.R
\name{simBatlen}
\alias{simBatlen}
\title{Calculates biomass at age for fishery independent surveys}
\usage{
simBatlen(Nlen, Flt, lsel, year)
}
\arguments{
\item{Nlen}{matrix with numbers at length}

\item{Flt}{number of fleets (2)}

\item{lsel}{selectivity-at-length matrix}

\item{year}{current year in simulation}
}
\description{
Calculates biomass at age for fishery independent surveys
}
\examples{
simBatlen()

}
\keyword{biomass-at-length}
