% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funs.R
\name{simIndex}
\alias{simIndex}
\title{Generates indices of abundance based on simulated catch}
\usage{
simIndex(dat.list, b)
}
\arguments{
\item{dat.list}{list with catchability and CPUE SE values}

\item{b}{vector of vulnerable biomass for each fleet}
}
\description{
Generates indices of abundance based on simulated catch
}
\keyword{abundance}
\keyword{index}
\keyword{of}
