% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_funs.R
\name{zatage}
\alias{zatage}
\title{Calcualte total mortality at age}
\usage{
zatage(dat.list, year, f.by.fleet)
}
\arguments{
\item{dat.list}{Contains natural mortality at age, matrix or df and selectivity-at-age matrix}

\item{year}{current year}

\item{f.by.fleet}{Fishing mortality by fleet for year y}
}
\description{
Calcualte total mortality at age
}
\keyword{mortality}
